Fix type error.

+++ src/EventListener/Dca/ContactProfileDcaListener.php
@@ -42,7 +42,7 @@ final class ContactProfileDcaListener
     public function toggleIcon(array $row, ?string $href, string $label, string $title, string $icon, string $attributes) : string
     {
         if (Input::get('tid') !== null && Input::get('tid') !== '') {
-            $this->toggleVisibility(Input::get('tid'), (Input::get('state') === '1'), (@func_get_arg(12) ?: null));
+            $this->toggleVisibility((int) Input::get('tid'), (Input::get('state') === '1'), (@func_get_arg(12) ?: null));
             Backend::redirect(Backend::getReferer());
         }
